pyproject.toml
[tool.poetry]
name = "encryptor"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
system = { path = "../system/", develop = false }
python = "^3.8"
pycryptodome = "3.14.1"
pycrypto = "2.6.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
